Rename the core service class
